+++ proto/cosmos/staking/v1beta1/tx.proto
@@ -3,11 +3,33 @@ package cosmos.staking.v1beta1;
 
 import "gogoproto/gogo.proto";
 import "cosmos/base/v1beta1/coin.proto";
+import "google/protobuf/timestamp.proto";
 import "cosmos/staking/v1beta1/staking.proto";
 
 option go_package = "github.com/cosmos/cosmos-sdk/x/staking/types";
 
-// MsgCreateValidator defines an SDK message for creating a new validator.
+// Msg defines the staking Msg service.
+service Msg {
+  // CreateValidator defines a method for creating a new validator.
+  rpc CreateValidator(MsgCreateValidator) returns (MsgCreateValidatorResponse);
+
+  // EditValidator defines a method for editing an existing validator.
+  rpc EditValidator(MsgEditValidator) returns (MsgEditValidatorResponse);
+
+  // Delegate defines a method for performing a delegation of coins
+  // from a delegator to a validator.
+  rpc Delegate(MsgDelegate) returns (MsgDelegateResponse);
+
+  // BeginRedelegate defines a method for performing a redelegation
+  // of coins from a delegator and source validator to a destination validator.
+  rpc BeginRedelegate(MsgBeginRedelegate) returns (MsgBeginRedelegateResponse);
+
+  // Undelegate defines a method for performing an undelegation from a
+  // delegate and a validator.
+  rpc Undelegate(MsgUndelegate) returns (MsgUndelegateResponse);
+}
+
+// MsgCreateValidator defines a SDK message for creating a new validator.
 message MsgCreateValidator {
   option (gogoproto.equal)           = false;
   option (gogoproto.goproto_getters) = false;
@@ -25,7 +47,10 @@ message MsgCreateValidator {
   cosmos.base.v1beta1.Coin value             = 7 [(gogoproto.nullable) = false];
 }
 
-// MsgEditValidator defines an SDK message for editing an existing validator.
+// MsgCreateValidatorResponse defines the Msg/CreateValidator response type.
+message MsgCreateValidatorResponse { }
+
+// MsgEditValidator defines a SDK message for editing an existing validator.
 message MsgEditValidator {
   option (gogoproto.equal)           = false;
   option (gogoproto.goproto_getters) = false;
@@ -48,7 +73,10 @@ message MsgEditValidator {
   ];
 }
 
-// MsgDelegate defines an SDK message for performing a delegation of coins
+// MsgEditValidatorResponse defines the Msg/EditValidator response type.
+message MsgEditValidatorResponse { }
+
+// MsgDelegate defines a SDK message for performing a delegation of coins
 // from a delegator to a validator.
 message MsgDelegate {
   option (gogoproto.equal)           = false;
@@ -59,7 +87,10 @@ message MsgDelegate {
   cosmos.base.v1beta1.Coin amount            = 3 [(gogoproto.nullable) = false];
 }
 
-// MsgBeginRedelegate defines an SDK message for performing a redelegation
+// MsgDelegateResponse defines the Msg/Delegate response type.
+message MsgDelegateResponse { }
+
+// MsgBeginRedelegate defines a SDK message for performing a redelegation
 // of coins from a delegator and source validator to a destination validator.
 message MsgBeginRedelegate {
   option (gogoproto.equal)           = false;
@@ -71,7 +102,12 @@ message MsgBeginRedelegate {
   cosmos.base.v1beta1.Coin amount                = 4 [(gogoproto.nullable) = false];
 }
 
-// MsgUndelegate defines an SDK message for performing an undelegation from a
+// MsgBeginRedelegateResponse defines the Msg/BeginRedelegate response type.
+message MsgBeginRedelegateResponse { 
+  google.protobuf.Timestamp completion_time = 1 [(gogoproto.nullable) = false, (gogoproto.stdtime) = true];
+}
+
+// MsgUndelegate defines a SDK message for performing an undelegation from a
 // delegate and a validator.
 message MsgUndelegate {
   option (gogoproto.equal)           = false;
@@ -80,4 +116,9 @@ message MsgUndelegate {
   string                   delegator_address = 1 [(gogoproto.moretags) = "yaml:\"delegator_address\""];
   string                   validator_address = 2 [(gogoproto.moretags) = "yaml:\"validator_address\""];
   cosmos.base.v1beta1.Coin amount            = 3 [(gogoproto.nullable) = false];
+}
+
+// MsgUndelegateResponse defines the Msg/Undelegate response type.
+message MsgUndelegateResponse { 
+  google.protobuf.Timestamp completion_time = 1 [(gogoproto.nullable) = false, (gogoproto.stdtime) = true];
 }
