+++ x/distribution/handler.go
@@ -9,10 +9,10 @@ import (
 )
 
 func NewHandler(k keeper.Keeper) sdk.Handler {
+	msgServer := keeper.NewMsgServerImpl(k)
+
 	return func(ctx sdk.Context, msg sdk.Msg) (*sdk.Result, error) {
 		ctx = ctx.WithEventManager(sdk.NewEventManager())
-
-		msgServer := keeper.NewMsgServerImpl(k)
 
 		switch msg := msg.(type) {
 		case *types.MsgSetWithdrawAddress:
